client/requirements.txt
requests
httpx
aiohttp

opentelemetry-api
opentelemetry-sdk

opentelemetry-exporter-otlp-proto-grpc

